More work on updater
Most of the updater code is now done. It will query an update file on the website which will then be used to see if there is an update. If there is it will then download the file for the current OS and Arch. Once the download is complete it verifies both the SHA1 checksum and the hash against a hardcoded signature. This prevents tampering of the download files across the network (for those security paranoid users!). I know the checksum isn't necassary, but I want to publish it in the updates.json file for users to manually compare their checksum, we might as well use it to quickly verify the checksum before doing the more computationally intensive verification.