Allow logging out of Trakt.tv from settings
Implemented a way of logging out of Trakt.tv from the settings dialog. It checks if App.Trakt.authenticated is true and then displays a button that lets you 'disconnect' your account. Fixes #832. This commit also contains a patch that extends the amount of time after a keyup event before it will check login details. This should fix some issues related to being unable to login to Trakt.tv accounts.