Fix nwjs build system on macOS
For osx, only 64-bit nwjs exists. There is no 32-bit nwjs anymore,
probably because Apple hasn't made 32-bit Macs for over a decade.

Related issue: https://github.com/nwjs/nw-builder/issues/363
"The 32 bit version works but only has old legacy versions and it is
not and probably won't be supported anymore." -- DblK @ nw-builder

So trying to build as osx32 would fail since it has no new nwjs versions.

In fact, building hasn't worked on OSX since nwjs 0.12.0 in early 2015,
which was when osx32 support was discontinued.

This fixes that problem. It now builds using the latest nwjs 64-bit,
which is the only available CPU architecture of nwjs for macOS.