Add support for .smi (or .sami) subtitles
You can now play .smi or .sami subtitles by selecting them from the
subtitles menu or drag-n-dropping them on the app when a movie is
playing. The subtitle will be converted to a .srt like the .txt, .ssa
and .ass subtitles since it's easier to have a single output point for
all subtitle files, but we could add unique markup features of the .smi
type (like colors and positioning) if it's desired in the future.