Add local subtitles support to chromecast streams
Before, only default subtitles embedded in the torrent link could be
played. Now, if the user drags a torrent file on the chromecast player,
we serve the local files to the subtitles server so that the chromecast
can access them like remote subtitles. Right now, only the subtitles
drag-and-dropping works, but the UI will come in a future update.