Revert to older delete folder function
In order to truly fix #1202 and #1280 it's mandatory to make some reverting changes. The old function was removed to make room for rimraf, which was used async. This would not have (90% of the time) removed all folders since it wouldn't wait for all the processes to end (hence async). So it was changed to sync, but a blocking error was thrown when you tried to close the app (it didn't actually wait for the processes to end).

In order to resolve this, I've reverted the  function back to the commit before 5cf7007 (this was back in 2015), where the favor for the async method took place. n
I also properly removed the WebTorrent when the streamer is killed, which allows the original function to still work.

Now it removes all the files from the temp folder, whilst still being able to exit.