Fix watch now btn not working on series episodes
tvdb_id value is sometimes null/undefined and a toString() is called on
it, which causes a crash. This happens if autoplay next episode is
selected in config.

This PR addresses this issue, checking of the value is defined before
calling toString(). In case it's undefined a blank string is used.