Clean subtitle file when 2 subtitle lines index are folowing each other
It was generating an "Error reading subtitle
timings, file seems to be corrupted" and subtitle
wasn't displayed.

issue is in subtitle syntax when it look like this:

```
3
2
00:00:59,142 --> 00:01:01,144
N'Y COMPTE PAS
```

Cleaning that (removing one index number) make the subtitle working.

How often it happen ? no idea. :/

More info in issue #1013