popcorn-time-beta cask now install latest sucessful dev build Removed specific version number and sha256 sum for lowering maintenance needs.
Now this cask will always install the latest sucessful dev build
for as long as the file in 'url' is always named
Popcorn-Time-0.4.4_osx64.zip

So, we lost sha256 checksum but we don't need to update this
file for every new dev build. I think it's better this way
unless someone want to update it everytime.