Better Seedbox stats updating
It was doing double the work (and more depending how many torrents you had loaded) with multiple update events with both `UpdateActiveTorrentView()` and `torrent.on('download', updateMetrics)` / `torrent.on('upload', updateMetrics)` active. Also at some point and after a few active torrents everything was getting desynced and the UI was behaving a bit crazy (and even more when the user was changing the selected torrent).

Now it just updates everything every 1000ms with the same timer, except the right pane which is additionally updated immediately after the user clicks/selects a specific torrent on the left side. (with the same function as well)