Fix max. active torrents number
'activeTorrentCount' includes the torrent being added, by the 'Watch now' or 'Download' function it doesn't matter, so the >= makes it e.g pause the 5th torrent if active torrents are set to 5 instead of the 6th. The same ofc happens if you change the default 5 to something else. It will pause number x instead of x+1. Changing >= to > fixes it.